:recycle: refactor: 사용자 대기열 추가 반환값 리팩토링
* Mono<ApiResponse<Long>> -> Mono<ApiResponse<Float>>

+++ queue/src/main/java/com/quit/queue/application/service/QueueService.java
@@ -20,7 +20,7 @@ import java.util.UUID;
 public class QueueService {
     private final ReactiveRedisTemplate<String, String> reactiveRedisTemplate;
 
-    public Mono<ApiResponse<Long>> addUserToQueueForStore(UUID storeId, Long userId) {
+    public Mono<ApiResponse<Float>> addUserToQueueForStore(UUID storeId, Long userId) {
         // TODO 권한 검증 추가
 
         String key = "queue:store:" + storeId + ":users";
@@ -37,9 +37,9 @@ public class QueueService {
                                         .map(ZSetOperations.TypedTuple::getScore)
                                         .switchIfEmpty(Mono.just(0.0))
                                         .flatMap(highestScore -> {
-                                            double newScore = (highestScore == 0.0) ? 1.0 : highestScore + 1;
+                                            float newScore = (highestScore == 0.0) ? 1.0f : (float) (highestScore + 1);
                                             return reactiveRedisTemplate.opsForZSet().add(key, userId.toString(), newScore)
-                                                    .then(Mono.just(ApiResponse.success((long) newScore)));
+                                                    .then(Mono.just(ApiResponse.success(newScore)));
                                         })
                                 );
                     }
