+++ queue/src/main/java/com/quit/queue/presentation/controller/QueueController.java
@@ -23,8 +23,8 @@ public class QueueController {
     }
 
     @PostMapping("/stores/{storeId}")
-    public Mono<ApiResponse<Long>> addUserToQueueForStore(@PathVariable UUID storeId,
-                                                          @RequestHeader(value = "X-User-Id") Long userId) {
+    public Mono<ApiResponse<Float>> addUserToQueueForStore(@PathVariable UUID storeId,
+                                                           @RequestHeader(value = "X-User-Id") Long userId) {
         return queueService.addUserToQueueForStore(storeId, userId);
     }
